log_ingestion_service/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
grpcio>=1.76.0,<2.0.0
protobuf>=5.0.0,<7.0.0
